--- CoinCap/ContentView.swift
@@ -1,24 +1,0 @@
-//
-//  ContentView.swift
-//  CoinCap
-//
-//  Created by Marina Lunts on 04.01.25.
-//
-
-import SwiftUI
-
-struct ContentView: View {
-    var body: some View {
-        VStack {
-            Image(systemName: "globe")
-                .imageScale(.large)
-                .foregroundStyle(.tint)
-            Text("Hello, world!")
-        }
-        .padding()
-    }
-}
-
-#Preview {
-    ContentView()
-}
