+++ CoinCap/Models/Asset+Preview.swift
@@ -5,24 +5,26 @@
 //  Created by Marina Lunts on 05.01.25.
 //
 
-import Foundation
-
 extension Asset {
-    static let preview1 = Self(
+    static let preview1 = Asset(
         id: "bitcoin",
-        iconURL: URL(string: "https://assets.coincap.io/assets/icons/btc@2x.png"),
-        name: "Bitcoin",
         symbol: "BTC",
-        price: "$97388.7",
-        percentageChange: "-0.54%"
+        name: "Bitcoin",
+        supply: "19000000",
+        marketCapUsd: "600000000000",
+        volumeUsd24Hr: "20000000000",
+        priceUsd: "30000",
+        changePercent24Hr: "-0.54"
     )
     
-    static let preview2 = Self(
+    static let preview2 = Asset(
         id: "ethereum",
-        iconURL: URL(string: "https://assets.coincap.io/assets/icons/eth@2x.png"),
-        name: "Ethereum",
         symbol: "ETH",
-        price: "$3603.31",
-        percentageChange: "-1.03%"
+        name: "Ethereum",
+        supply: "120000000",
+        marketCapUsd: "200000000000",
+        volumeUsd24Hr: "10000000000",
+        priceUsd: "1800",
+        changePercent24Hr: "0.44"
     )
 }
